// web/lib/api/generated/model/enteredUnitEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `BASE` - Base unit
 * * `PACK` - Pack
 */
export type EnteredUnitEnum = typeof EnteredUnitEnum[keyof typeof EnteredUnitEnum];


export const EnteredUnitEnum = {
  BASE: 'BASE',
  PACK: 'PACK',
} as const;