// web/lib/api/generated/model/receiptLineInputRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */
import type { EnteredUnitEnum } from './enteredUnitEnum';

export interface ReceiptLineInputRequest {
  /** @nullable */
  id?: string | null;
  product_id: string;
  entered_unit?: EnteredUnitEnum;
  /** @pattern ^-?\d{0,11}(?:\.\d{0,3})?$ */
  entered_qty: string;
  /**
     * @nullable
     * @pattern ^-?\d{0,10}(?:\.\d{0,4})?$
     */
  entered_cost?: string | null;
}